requirements.txt
# Core dependencies
python-telegram-bot
instagrapi
pydantic
pydantic-settings
python-dotenv

# Utilities (keep these)
pyotp  # Still useful for 2FA
requests
aiofiles
tabulate
pillow

# Development dependencies (unchanged)
black
isort
mypy
pylint
pytest
pytest-asyncio